+++ packages/hooks/src/index.js
@@ -1,2 +1,3 @@
 export * from './useEthscanBalance'
 export * from './useEthscanTokensBalance'
+export * from './useTokensPrice'